Bug: #1228 Bug: #288
Upon further review, NHibernate requires a specific signed version of the log4net.dll, or else it won't function.  Reverted back to the previous version of the DLL, and backed out changes to the Logging.config files.  Note that I stored the changed log4net.dll file in ReferencedAssemblies/log4net.dll, in case we can use it in the future.